fix(init) : add cleanup venv path in PATH on `activate_venv()`
1) Problem

Previous code cannot remove existing path which has `.venv` directory.
It always add .venv directory at first. it has an limit that the PATH
will increase.
So It need to edit a file in only one project per neovim instance.
Plus, When I change focus to other file which is not in python venv
project, It need to go out from virtual environment with making
$VIRTUAL_ENV as nil. but it doesn't. the PATH is remained.

2) Solution

Make Clean up process when activate_venv() function. If It is executed
in current file directory which doesn't have .venv directory, the PATH
that has .venv will be removed in $PATH.

If there are .venv directory in pwd, add .venv path in PATH.
It need to check that user change pwd to project root always.